pyproject.toml
[project]
name = "openclaw-hobby"
version = "0.1.0"
description = "基于 OpenClaw 的自主兴趣内容策展系统"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
dependencies = [
    "requests>=2.28",
    "feedparser>=6.0",
]

[project.optional-dependencies]
youtube = ["google-auth-oauthlib", "google-api-python-client"]
feishu = []  # feishu only needs requests (already in base deps)
dev = ["pytest>=7.0"]
all = ["google-auth-oauthlib", "google-api-python-client"]

[project.scripts]
hobee = "hobee.cli:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["hobee*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
